Fix type conversion

+++ include/child.h
@@ -27,6 +27,6 @@ int treat_client (int, struct sockaddr const*, int, config const*);
 int connect_one (struct addrinfo const*);
 int connect_host (char const *, int);
 int start_proxying (service const *, int *, int *);
-int forward_data (int, int, char *, unsigned int, int*, int*);
+int forward_data (int, int, char *, unsigned int, ssize_t*, ssize_t*);
 
 #endif
